+++ app/com/lynxanalytics/biggraph/controllers/SQLController.scala
@@ -550,7 +550,6 @@ class SQLController(val env: BigGraphEnvironment) {
   def exportSQLQueryToTable(
     user: serving.User, request: SQLExportToTableRequest) = async[Unit] {
     val df = request.dfSpec.createDataFrame(user, SQLController.defaultContext(user))
-
     SQLController.saveTable(
       df, s"From ${request.dfSpec.project} by running ${request.dfSpec.sql}",
       user, request.table, request.privacy, request.overwrite,
@@ -655,6 +654,7 @@ object SQLController {
     assertAccessAndGetTableEntry(user, tableName, privacy)
     val table = TableImport.importDataFrameAsync(df)
     val entry = assertAccessAndGetTableEntry(user, tableName, privacy)
+    assert(!entry.exists || overwrite, "file-already-exists-confirm-overwrite")
     val checkpoint = table.saveAsCheckpoint(notes)
     if (overwrite) entry.remove()
     val frame = entry.asNewTableFrame(checkpoint)
@@ -668,6 +668,7 @@ object SQLController {
       implicit metaManager: MetaGraphManager,
       dataManager: DataManager) = {
     val entry = assertAccessAndGetTableEntry(user, name, privacy)
+    assert(!entry.exists || overwrite, "file-already-exists-confirm-overwrite")
     if (overwrite) entry.remove()
     val view = entry.asNewViewFrame(recipe, notes)
     FEOption.titledCheckpoint(view.checkpoint, name, s"|${name}")